# Non-decoding state view