--- a/bin/convert/main.ts
+++ b/bin/convert/main.ts
@@ -101,18 +101,28 @@ function dumpOutput(
       if (type.encode === undefined) {
         throw new Error(`${type.name} does not support encoding to JAM codec.`);
       }
-      const encoded = Encoder.encodeObject(type.encode, data, spec);
+      let encoded: BytesBlob;
+      if (typeof type.encode === "function") {
+        encoded = Encoder.encodeObject(type.encode(spec), data, spec);
+      } else {
+        encoded = Encoder.encodeObject(type.encode, data, spec);
+      }
       dump(`${encoded}`);
       return;
     }
     case OutputFormat.Bin: {
-      if (type.encode === undefined) {
-        throw new Error(`${type.name} does not support encoding to JAM codec.`);
-      }
       if (destination === null) {
         throw new Error(`${OutputFormat.Bin} requires destination file.`);
       }
-      const encoded = Encoder.encodeObject(type.encode, data, spec);
+      if (type.encode === undefined) {
+        throw new Error(`${type.name} does not support encoding to JAM codec.`);
+      }
+      let encoded: BytesBlob;
+      if (typeof type.encode === "function") {
+        encoded = Encoder.encodeObject(type.encode(spec), data, spec);
+      } else {
+        encoded = Encoder.encodeObject(type.encode, data, spec);
+      }
       dump(encoded.raw);
       return;
     }
@@ -260,7 +270,11 @@ async function loadAndProcessDataFile(
     if (decodeType.decode === undefined) {
       throw new Error(`${decodeType.name} does not support decoding from binary data.`);
     }
-    data = Decoder.decodeObject(decodeType.decode, input.data, spec);
+    if (typeof decodeType.decode === "function") {
+      data = Decoder.decodeObject(decodeType.decode(spec), input.data, spec);
+    } else {
+      data = Decoder.decodeObject(decodeType.decode, input.data, spec);
+    }
   } else if (input.type === "json") {
     if (decodeType.json === undefined) {
       throw new Error(`${decodeType.name} does not support reading from JSON.`);