--- a/packages/jam/block/gp-constants.ts
+++ b/packages/jam/block/gp-constants.ts
@@ -61,11 +61,3 @@ export const W_X = 3_072;
  */
 export const MAX_REPORT_DEPENDENCIES = 8;
 export type MAX_REPORT_DEPENDENCIES = typeof MAX_REPORT_DEPENDENCIES;
-
-/** `Q`: Size of the authorization queue. */
-export const AUTHORIZATION_QUEUE_SIZE = Q;
-export type AUTHORIZATION_QUEUE_SIZE = typeof AUTHORIZATION_QUEUE_SIZE;
-
-/** `O`: Maximal authorization pool size. */
-export const MAX_AUTH_POOL_SIZE = O;
-export type MAX_AUTH_POOL_SIZE = typeof MAX_AUTH_POOL_SIZE;