--- a/packages/jam/block/utils.ts
+++ b/packages/jam/block/utils.ts
@@ -1,5 +1,5 @@
-import {Decoder, Descriptor, Encoder} from "@typeberry/codec";
-import {ChainSpec} from "@typeberry/config";
+import { Decoder, type Descriptor, Encoder } from "@typeberry/codec";
+import type { ChainSpec } from "@typeberry/config";
 
 /**
  * Take an input data and re-encode that data as view.
@@ -7,12 +7,7 @@ import {ChainSpec} from "@typeberry/config";
  * NOTE: this function should NEVER be used in any production code,
  * it's only a test helper.
  */
-export function reencodeAsView<T, V>(
-  codec: Descriptor<T, V>,
-  object: T,
-  chainSpec?: ChainSpec,
-): V {
+export function reencodeAsView<T, V>(codec: Descriptor<T, V>, object: T, chainSpec?: ChainSpec): V {
   const encoded = Encoder.encodeObject(codec, object, chainSpec);
   return Decoder.decodeObject(codec.View, encoded, chainSpec);
 }
-