--- a/packages/jam/state-json/dump.ts
+++ b/packages/jam/state-json/dump.ts
@@ -1,5 +1,4 @@
 import { type EntropyHash, type PerEpochBlock, tryAsPerEpochBlock, tryAsServiceGas } from "@typeberry/block";
-import { AUTHORIZATION_QUEUE_SIZE, MAX_AUTH_POOL_SIZE } from "@typeberry/block/gp-constants.js";
 import type { AuthorizerHash, WorkPackageHash } from "@typeberry/block/refine-context.js";
 import { fromJson } from "@typeberry/block-json";
 import { Bytes } from "@typeberry/bytes";
@@ -9,9 +8,11 @@ import { BANDERSNATCH_RING_ROOT_BYTES } from "@typeberry/crypto/bandersnatch.js"
 import { json } from "@typeberry/json-parser";
 import {
   type AccumulationOutput,
+  AUTHORIZATION_QUEUE_SIZE,
   accumulationOutputComparator,
   type InMemoryService,
   InMemoryState,
+  MAX_AUTH_POOL_SIZE,
   PrivilegedServices,
   RecentBlocks,
   type State,