--- a/packages/jam/state-json/not-yet-accumulated.ts
+++ b/packages/jam/state-json/not-yet-accumulated.ts
@@ -1,6 +1,6 @@
 import { fromJson, workReportFromJson } from "@typeberry/block-json";
 import { json } from "@typeberry/json-parser";
-import { NotYetAccumulatedReport } from "@typeberry/state/not-yet-accumulated.js";
+import { NotYetAccumulatedReport } from "@typeberry/state/accumulation-queue.js";
 
 export const notYetAccumulatedFromJson = json.object<NotYetAccumulatedReport>(
   {