--- a/packages/jam/state/assurances.ts
+++ b/packages/jam/state/assurances.ts
@@ -1,8 +1,8 @@
 import type { TimeSlot } from "@typeberry/block";
 import { WorkReport } from "@typeberry/block/work-report.js";
-import { type CodecRecord, codec, DescribedBy } from "@typeberry/codec";
+import { type CodecRecord, codec, type DescribedBy } from "@typeberry/codec";
 import { WithDebug } from "@typeberry/utils";
-import {codecPerCore} from "./common.js";
+import { codecPerCore } from "./common.js";
 
 /**
  * Assignment of particular work report to a core.