and some other performance fixes.

This PR introduces a non-decoding way to inspect the state. It's integrated via `view()` method on the state object (both `InMemoryState` and `SerializedState`). However accessing views with `InMemoryState` is not really expected to be faster, because we need to re-encode the objects.

Using views is much more cumbersome and requires a very specific usage pattern to actually be faster in benchmarks, so it does not mean we have to re-write the entire state transition to use views. Rather I think we should be introducing them carefully when it's actually beneficial as shown by profiler.

The usage is shown only in one place currently - `reports` transition, when we avoid accessing the entire `authPools` object.

Some additional performance improvements smuggled in this PR:
- [x] Introduce cache for decoded objects in state
- [x] fix sequence views dependent on context (they were incorrect due to codec being dependent on context, see also #306)
- [x] clean up too much data being passed to safrole seal
- [x] optimize extrinsic hash calculation (less allocation)
- [x] clean up recent blocks compatibility stuff


<details>
<summary>Performance stats for this PR </summary>

<h2 class="text-xl font-bold text-text-100 mt-1 -mb-0.5">fallback</h2>

Metric | PR (μs) | main (μs) | Difference (μs) | % Change
-- | -- | -- | -- | --
min | 15,504.6 | 18,574.1 | -3,069.5 | -16.5%
max | 39,314.0 | 34,505.1 | +4,808.9 | +13.9%
mean | 17,065.0 | 20,519.9 | -3,454.9 | -16.8%
median | 16,815.8 | 20,672.3 | -3,856.5 | -18.7%
stdDev | 1,522.0 | 1,354.7 | +167.3 | +12.4%
p90 | 18,631.1 | 21,827.9 | -3,196.8 | -14.6%
p99 | 20,920.0 | 24,961.5 | -4,041.5 | -16.2%

<h2 class="text-xl font-bold text-text-100 mt-1 -mb-0.5">safrole</h2>

Metric | PR (μs) | main (μs) | Difference (μs) | % Change
-- | -- | -- | -- | --
min | 15,484.8 | 18,552.3 | -3,067.5 | -16.5%
max | 62,988.7 | 89,843.4 | -26,854.7 | -29.9%
mean | 31,097.4 | 44,453.5 | -13,356.1 | -30.0%
median | 29,773.5 | 42,179.3 | -12,405.8 | -29.4%
stdDev | 14,142.3 | 23,803.1 | -9,660.8 | -40.6%
p90 | 53,858.4 | 83,655.7 | -29,797.3 | -35.6%
p99 | 55,904.3 | 85,985.2 | -30,080.9 | -35.0%

<h2 class="text-xl font-bold text-text-100 mt-1 -mb-0.5">storage</h2>

Metric | PR (μs) | main (μs) | Difference (μs) | % Change
-- | -- | -- | -- | --
min | 16,148.8 | 16,216.6 | -67.8 | -0.4%
max | 321,272.4 | 339,053.7 | -17,781.3 | -5.2%
mean | 85,305.1 | 89,160.7 | -3,855.6 | -4.3%
median | 52,160.8 | 54,508.9 | -2,348.1 | -4.3%
stdDev | 74,996.8 | 77,439.0 | -2,442.2 | -3.2%
p90 | 203,431.5 | 204,565.2 | -1,133.7 | -0.6%
p99 | 298,292.6 | 307,485.6 | -9,193.0 | -3.0%

<h2 class="text-xl font-bold text-text-100 mt-1 -mb-0.5">storage_light</h2>

Metric | PR (μs) | main (μs) | Difference (μs) | % Change
-- | -- | -- | -- | --
min | 15,948.9 | 16,113.5 | -164.6 | -1.0%
max | 121,333.8 | 182,211.6 | -60,877.8 | -33.4%
mean | 39,762.3 | 40,474.9 | -712.6 | -1.8%
median | 32,750.9 | 33,538.5 | -787.6 | -2.3%
stdDev | 25,927.8 | 26,283.8 | -356.0 | -1.4%
p90 | 80,161.7 | 82,011.9 | -1,850.2 | -2.3%
p99 | 107,618.6 | 110,393.1 | -2,774.5 | -2.5%


</details>